+++ .changeset/nine-kings-divide.md
@@ -1,0 +1,5 @@
+---
+"landscape-ui": minor
+---
+
+Update instance modals
