+++ src/features/instances/components/RestartModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./RestartModal";
