+++ src/features/instances/components/ShutDownModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ShutDownModal";
