+++ src/features/instances/index.ts
@@ -23,3 +23,5 @@ export { default as RegenerateRecoveryKeyModal } from "./components/RegenerateRe
 export { default as ViewRecoveryKeyModal } from "./components/ViewRecoveryKeyModal";
 export type { Status, RecoveryKey, DistributionUpgradeTarget } from "./types";
 export { default as useInstanceSearchHelpTerms } from "./components/InstancesHeader/hooks/useInstanceSearchHelpTerms";
+export { default as RestartModal } from "./components/RestartModal";
+export { default as ShutDownModal } from "./components/ShutDownModal";
