+++ src/pages/dashboard/instances/[single]/tabs/info/InfoPanel/InfoPanel.tsx
@@ -20,10 +20,9 @@ import {
   RecoveryKeyStatus,
   RegenerateRecoveryKeyModal,
   useGetRecoveryKey,
-  useRestartInstances,
   useSanitizeInstance,
-  useShutDownInstances,
   ViewRecoveryKeyModal,
+  ShutDownModal,
 } from "@/features/instances";
 import {
   WslInstanceReinstallModal,
@@ -42,29 +41,23 @@ import type {
   WslInstance,
 } from "@/types/Instance";
 import { hasOneItem } from "@/utils/_helpers";
-import { getFormikError } from "@/utils/formikErrors";
 import {
   Button,
-  CheckboxInput,
   ConfirmationModal,
-  Form,
   Icon,
   ICONS,
-  Input,
   Tooltip,
 } from "@canonical/react-components";
 import classNames from "classnames";
-import { useFormik } from "formik";
 import moment from "moment";
 import type { FC } from "react";
 import { Fragment, lazy, Suspense } from "react";
 import { useNavigate } from "react-router";
 import { useBoolean } from "usehooks-ts";
 import ProfileLink from "../ProfileLink";
-import { INITIAL_VALUES, VALIDATION_SCHEMA } from "./constants";
 import { getInstanceKeyForRemount } from "./helpers";
 import classes from "./InfoPanel.module.scss";
-import type { ModalConfirmationFormProps } from "./types";
+import { RestartModal } from "@/features/instances";
 
 const EditInstance = lazy(
   async () =>
@@ -98,7 +91,6 @@ const InfoPanel: FC<InfoPanelProps> = ({ instance }) => {
     { id: instance.employee_id ?? 0 },
     { enabled: !!instance.employee_id },
   );
-  const { restartInstances, isRestartingInstances } = useRestartInstances();
   const { recoveryKey, recoveryKeyActivityStatus, isRecoveryKeyFetched } =
     useGetRecoveryKey(instance.id);
   const isRecoveryKeyGenerationActivityInProgress =
@@ -110,7 +102,6 @@ const InfoPanel: FC<InfoPanelProps> = ({ instance }) => {
     );
   const { getAccessGroupQuery } = useRoles();
   const { sanitizeInstance, isSanitizingInstance } = useSanitizeInstance();
-  const { shutDownInstances, isShuttingDownInstances } = useShutDownInstances();
   const { disassociateEmployeeFromInstance, isDisassociating } =
     useDisassociateEmployeeFromInstance();
 
@@ -176,59 +167,6 @@ const InfoPanel: FC<InfoPanelProps> = ({ instance }) => {
 
   const { data: getAccessGroupQueryResult, isPending: isGettingAccessGroups } =
     getAccessGroupQuery();
-
-  const handleSubmit = async (values: ModalConfirmationFormProps) => {
-    if (!values.action) {
-      return;
-    }
-
-    const valuesToSubmit = {
-      computer_ids: [instance.id],
-      deliver_after: values.deliverImmediately
-        ? undefined
-        : `${values.deliver_after}:00Z`,
-    };
-
-    try {
-      const { data: activity } =
-        values.action === "reboot"
-          ? await restartInstances(valuesToSubmit)
-          : await shutDownInstances(valuesToSubmit);
-
-      const notificationVerb =
-        values.action === "reboot" ? "restarted" : "shut down";
-
-      notify.success({
-        title: `You queued "${instance.title}" to be ${notificationVerb}`,
-        message: `Instance "${instance.title}" will be ${notificationVerb} and is queued in Activities`,
-        actions: [
-          {
-            label: "View details",
-            onClick: () => {
-              openActivityDetails(activity);
-            },
-          },
-        ],
-      });
-
-      switch (values.action) {
-        case "reboot":
-          closeRestartModal();
-          break;
-        case "shutdown":
-          closeShutDownModal();
-          break;
-      }
-    } catch (error) {
-      debug(error);
-    }
-  };
-
-  const formik = useFormik({
-    initialValues: INITIAL_VALUES,
-    validationSchema: VALIDATION_SCHEMA,
-    onSubmit: handleSubmit,
-  });
 
   if (isGettingAccessGroups || isGettingEmployee) {
     return <LoadingState />;
@@ -316,11 +254,6 @@ const InfoPanel: FC<InfoPanelProps> = ({ instance }) => {
     } catch (error) {
       debug(error);
     }
-  };
-
-  const handleFormSubmit = async (action: "reboot" | "shutdown") => {
-    await formik.setFieldValue("action", action);
-    formik.handleSubmit();
   };
 
   const goBack = () => {
@@ -613,65 +546,17 @@ const InfoPanel: FC<InfoPanelProps> = ({ instance }) => {
         </Blocks.Item>
       </Blocks>
 
-      {isRestartModalOpen && (
-        <ConfirmationModal
-          close={closeRestartModal}
-          title="Restart instance"
-          confirmButtonLabel="Restart"
-          confirmButtonAppearance="positive"
-          confirmButtonDisabled={isRestartingInstances}
-          confirmButtonLoading={isRestartingInstances}
-          onConfirm={async () => handleFormSubmit("reboot")}
-        >
-          <Form onSubmit={async () => handleFormSubmit("reboot")} noValidate>
-            <CheckboxInput
-              label="Deliver as soon as possible"
-              {...formik.getFieldProps("deliverImmediately")}
-              checked={formik.values.deliverImmediately}
-            />
-            <Input
-              type="datetime-local"
-              label="Schedule time"
-              labelClassName="u-off-screen"
-              placeholder="Scheduled time"
-              {...formik.getFieldProps("deliver_after")}
-              disabled={formik.values.deliverImmediately}
-              error={getFormikError(formik, "deliver_after")}
-            />
-            <p>This will restart &quot;{instance.title}&quot; instance.</p>
-          </Form>
-        </ConfirmationModal>
-      )}
+      <RestartModal
+        instances={[instance]}
+        isOpen={isRestartModalOpen}
+        close={closeRestartModal}
+      />
 
-      {isShutDownModalOpen && (
-        <ConfirmationModal
-          close={closeShutDownModal}
-          title="Shut down instance"
-          confirmButtonLabel="Shut down"
-          confirmButtonAppearance="positive"
-          confirmButtonDisabled={isShuttingDownInstances}
-          confirmButtonLoading={isShuttingDownInstances}
-          onConfirm={async () => handleFormSubmit("shutdown")}
-        >
-          <Form onSubmit={async () => handleFormSubmit("shutdown")} noValidate>
-            <CheckboxInput
-              label="Deliver as soon as possible"
-              {...formik.getFieldProps("deliverImmediately")}
-              checked={formik.values.deliverImmediately}
-            />
-            <Input
-              type="datetime-local"
-              label="Schedule time"
-              labelClassName="u-off-screen"
-              placeholder="Scheduled time"
-              {...formik.getFieldProps("deliver_after")}
-              disabled={formik.values.deliverImmediately}
-              error={getFormikError(formik, "deliver_after")}
-            />
-            <p>This will shut down &quot;{instance.title}&quot; instance.</p>
-          </Form>
-        </ConfirmationModal>
-      )}
+      <ShutDownModal
+        instances={[instance]}
+        isOpen={isShutDownModalOpen}
+        close={closeShutDownModal}
+      />
 
       {disassociateModalOpen && (
         <ConfirmationModal
